src/agent/alfred_agent/requirements.txt
google-adk==1.14.0
google-cloud-firestore>=2.19.0
langchain-community==0.3.27
wikipedia==1.4.0
fastapi
uvicorn
requests
python-dotenv
mcp